Fundamental Kotlin - Mapped properties.
